// src/generated/model/gateadresseFrontend.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface GateadresseFrontend {
    landkode?: string;
    kommunenummer?: string;
    adresselinjer?: string[];
    bolignummer?: string;
    postnummer?: string;
    poststed?: string;
    gatenavn?: string;
    husnummer?: string;
    husbokstav?: string;
}